print first 98 fibo nums. new test 5

+++ functions_nested_loops/104-fibonacci.c
@@ -7,13 +7,13 @@
 int main(void)
 {
 	int max = 98, term;
-	long int fib, fib_term1 = 1, fib_term2 = 2;
+	long long int fib, fib_term1 = 1, fib_term2 = 2;
 
 	printf("%ld, %ld, ", fib_term1, fib_term2);
 	for (term = 0; term < max; term++)
 	{
 		fib = fib_term1 + fib_term2;
-		printf("%ld", fib);
+		printf("%lld", fib);
 		if (term + 1 < max)
 			printf(", ");
 		fib_term1 = fib_term2;
